requirements.txt
Flask==2.3.2
pymongo==4.6.1
python-dotenv==1.0.1

